+++ util/VCode.cpp
@@ -287,7 +287,7 @@ static bool vcode_crc_selftest( void ) {
 #if !defined(HWCRC_U64)
     if (use_hw) { return true; }
 #endif
-    constexpr uint32_t testcnt = 6;
+    constexpr size_t testcnt = 6;
     uint8_t offsets[testcnt]   = { 0x01, 0x29, 0x51, 0x79, 0xa1, 0xc9 };
     uint32_t crcs[testcnt]     = {
         0x0e2c157f, 0xe980ebf6, 0xde74bded,
@@ -295,7 +295,7 @@ static bool vcode_crc_selftest( void ) {
     };
     uint32_t crc;
 
-    for (int i = 0; i < testcnt; i++) {
+    for (size_t i = 0; i < testcnt; i++) {
         crc = vcode_crc_selftest_40<use_hw, true>( offsets[i]);
         if (crc != crcs[i]) { return false; }
